--- src/main/java/Calculator.java
@@ -1,9 +1,0 @@
-public class Calculator {
-    int multiply(int a, int b) {
-        return a * b;
-    }
-
-    int multiply(int a, int b, int c) {
-        return a * b * c;
-    }
-}
